// lingxi-wx-frontend/src/types/types.ts
export type MessageType = {
    /** 发送者信息 */
    fromUser: MsgUserType;
    /** 消息主体 */
    message: MsgType;
    /** 发送时间 */
    sendTime: string;
    /** 时间段（可选） */
    timeBlock?: string;
    /** 是否加载中 */
    loading?: boolean;
};

/**
 * 消息中用户信息
 */
export type MsgUserType = {
    /** 用户ID */
    uid: string;
    /** 用户名 */
    username: string;
    /** 头像 */
    avatar: string;
    /** 归属地 */
    locPlace: string;
    /** 徽章 */
    badge?: {
        /** 徽章地址 */
        img: string;
        /** 描述 */
        describe: string; // 描述
    };
};

/**
 * 消息互动信息
 */
export type MessageMarkType = {
    /** 点赞 */
    userLike: number;
    /** 举报 */
    userDislike: number;
    /** 点赞数 */
    likeCount: number;
    /** 举报数 */
    dislikeCount: number;
};

/** 图片消息体 */
export type ImageBody = {
    size: number;
    url: string;
    width: number;
    height: number;
};
/** 语音消息体 */
export type VoiceBody = {
    size: number;
    second: number;
    url: string;
};
/** 视频 */
export type VideoBody = {
    size: number;
    url: string;
    thumbSize?: number;
    thumbWidth?: number;
    thumbHeight?: number;
    thumbUrl?: string;
};
/** 文件消息体 */
export type FileBody = {
    size: number;
    fileName: string;
    url: string;
};
/** 文本消息体 */
export type TextBody = {
    /** 消息内容 */
    content: string;
    /** 回复 */
    reply: ReplyType;
    /**
     * 消息链接映射
     */
    urlContentMap: Record<
        string,
        {
            title: string;
            description: string;
            image: string;
        }
    >;
};
/** 表情消息 */
export type EmojiBody = {
    url: string;
};

/**
 * 消息内容
 */
export type MsgType = {
    /** 消息ID */
    id: number;
    /**  房间 ID */
    roomId: number;
    /** 消息类型 */
    type: MsgEnum;
    /** 动态消息体-`根据消息类型变化` */
    body: TextBody | ImageBody | VoiceBody | VideoBody | FileBody | EmojiBody | any;
    /** 发送时间戳 */
    sendTime: number;
    /** 消息互动信息 */
    messageMark: MessageMarkType;
};
// ...
export enum IsAllUserEnum {
    /** 0否 */
    Not,
    /** 1是 */
    Yes
}

/** 会话列表项 */
export type SessionItem = {
    /** 房间最后活跃时间(用来排序) */
    activeTime: number;
    /** 会话头像 */
    avatar: string;
    /** 是否全员展示的会话 0否 1是 */
    hot_Flag: IsAllUserEnum;
    /** 会话名称 */
    name: string;
    /** 房间id */
    roomId: number;
    /** 最新消息 */
    text: string;
    /** 房间类型 1群聊 2单聊 */
    type: RoomTypeEnum;
    /** 未读数 */
    unreadCount: number;
};